src/requirements.txt
Flask==1.1.2
numpy==1.20.1
scikit_image==0.18.1
joblib==1.0.1
flask_cors==3.0.10
mysql_connector_repackaged==0.3.1
Pillow==8.3.2